+++ __tests__/services/world_engine.service.spec.ts
@@ -85,6 +85,48 @@ describe('load scene', () => {
     expect(mockLoadScene.mock.calls[0][0].client).toEqual(sceneClient);
   });
 
+  test("should use default user id if it's not provided", async () => {
+    const user = { fullName: 'Full Name' };
+    const mockLoadScene = jest.fn(
+      (_req: LoadSceneRequest, _initReq?: fm.InitReq) => {
+        return Promise.resolve({ agents });
+      },
+    );
+
+    WorldEngine.LoadScene = mockLoadScene;
+
+    await client.loadScene({
+      config: {
+        capabilities,
+        connection: {
+          gateway: { hostname: 'examples.com', ssl: true },
+        },
+      },
+      name: SCENE,
+      session,
+      user,
+    });
+
+    const sentUser = mockLoadScene.mock.calls[0][0].user;
+
+    expect(sentUser.name).toEqual(user.fullName);
+    expect(sentUser.id.length).not.toEqual(0);
+
+    await client.loadScene({
+      config: {
+        capabilities,
+        connection: {
+          gateway: { hostname: 'examples.com', ssl: true },
+        },
+      },
+      name: SCENE,
+      session,
+      user,
+    });
+
+    expect(sentUser.id).toEqual(mockLoadScene.mock.calls[1][0].user.id);
+  });
+
   test('should use http url is ssl option is not provided', async () => {
     const mockLoadScene = jest.fn(
       (_req: LoadSceneRequest, _initReq?: fm.InitReq) => {
@@ -105,6 +147,80 @@ describe('load scene', () => {
       name: SCENE,
       session,
       user,
+    });
+  });
+
+  test('should use provided provided user id', async () => {
+    const mockLoadScene = jest.fn(
+      (_req: LoadSceneRequest, _initReq?: fm.InitReq) => {
+        return Promise.resolve({ agents });
+      },
+    );
+
+    WorldEngine.LoadScene = mockLoadScene;
+
+    await client.loadScene({
+      config: {
+        capabilities,
+        connection: {
+          gateway: { hostname: 'examples.com', ssl: true },
+        },
+      },
+      name: SCENE,
+      session,
+      user: { id: user.id },
+    });
+    expect(mockLoadScene.mock.calls[0][0].user).toEqual({ id: user.id });
+  });
+
+  test('should use provided provided user name', async () => {
+    const mockLoadScene = jest.fn(
+      (_req: LoadSceneRequest, _initReq?: fm.InitReq) => {
+        return Promise.resolve({ agents });
+      },
+    );
+
+    WorldEngine.LoadScene = mockLoadScene;
+
+    await client.loadScene({
+      config: {
+        capabilities,
+        connection: {
+          gateway: { hostname: 'examples.com', ssl: true },
+        },
+      },
+      name: SCENE,
+      session,
+      user: { fullName: user.fullName },
+    });
+    expect(mockLoadScene.mock.calls[0][0].user.name).toEqual(user.fullName);
+  });
+
+  test('should use provided provided user profile', async () => {
+    const mockLoadScene = jest.fn(
+      (_req: LoadSceneRequest, _initReq?: fm.InitReq) => {
+        return Promise.resolve({ agents });
+      },
+    );
+
+    WorldEngine.LoadScene = mockLoadScene;
+
+    await client.loadScene({
+      config: {
+        capabilities,
+        connection: {
+          gateway: { hostname: 'examples.com', ssl: true },
+        },
+      },
+      name: SCENE,
+      session,
+      user: { profile: user.profile },
+    });
+    expect(
+      mockLoadScene.mock.calls[0][0].userSettings.playerProfile.fields[0],
+    ).toEqual({
+      fieldId: user.profile.fields[0].id,
+      fieldValue: user.profile.fields[0].value,
     });
   });
 });